typo/style migration and pre-commit for mrp_workcenter_workorder_link
