+++ mrp_workcenter_workorder_link/__manifest__.py
@@ -1,0 +1,21 @@
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
+
+{
+    "name": "MRP Workcenter Workorder Link",
+    "version": "14.0.1.0.0",
+    "author": "Akretion,Odoo Community Association (OCA)",
+    "summary": "Switch easily between Work Centers and Work Orders",
+    "maintener": "Akretion",
+    "category": "Manufacturing",
+    "depends": [
+        "mrp",
+    ],
+    "website": "https://github.com/OCA/manufacture",
+    "data": [
+        "views/mrp_workcenter_views.xml",
+        "views/mrp_workorder_views.xml",
+    ],
+    "license": "AGPL-3",
+    "installable": True,
+    "application": False,
+}
