+++ mrp_workcenter_workorder_link/models/__init__.py
@@ -1,21 +1,4 @@
-##############################################################################
-#
-#    Copyright (C) All Rights Reserved 2015 Akretion
-#    @author David BEAL <david.beal@akretion.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-###############################################################################
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
-from . import mrp
+from . import mrp_workcenter
+from . import mrp_workorder
