--- mrp_workcenter_workorder_link/static/description/index.html
@@ -1,21 +1,0 @@
-
-<section class="oe_container oe_website_only">
-
-<div class="oe_row oe_more_spaced">
-    <h3 class="oe_slogan">Workcenter Workorder Link</h3>
-    <h3 class="oe_slogan"><a href="http://odoo-community.org/">An OCA module</a> : validated by Odoo community experts</h3>
-    <h4 class="oe_slogan"><a href="http://www.akretion.com/fr">By Akretion</a></h4>
-<div class="oe_span6">
-
-<h2>Workcenter Workorder Link</h2>
-
-<table>
-    <tr><td><h3>Work Orders in activity on Workcenter screen</h3><div class="oe_row_img oe_centered"><img class="oe_picture" src="workc.png"></div><br/><br/></td></tr>
-    <tr><td><h3>Button to directly jump into Work Orders of the workcenter</h3><div class="oe_row_img oe_centered"><img class="oe_picture" src="tree.png"></div><br/><br/></td></tr>
-</table>
-<div class="oe_row oe_centeralign oe_more_space">
-    <a href="http://www.akretion.com/contact" class="oe_button oe_big">Contact us</a><br>
-    <a href="https://github.com/OCA/manufacture/tree/7.0" class="oe_button oe_big">Access to the code</a>
-</div>
-
-</section>
