+++ mrp_workcenter_workorder_link/tests/test_action.py
@@ -1,4 +1,4 @@
-from openerp.tests.common import TransactionCase
+from odoo.tests.common import TransactionCase
 
 
 class ExecuteButton(TransactionCase):
